added new endpoint and updated schema for new endpoints

+++ src/main/java/com/walmart/aex/sp/service/impl/DeptAdminRuleServiceImpl.java
@@ -2,6 +2,7 @@ package com.walmart.aex.sp.service.impl;
 
 import com.walmart.aex.sp.dto.deptadminrule.DeptAdminRuleRequest;
 import com.walmart.aex.sp.dto.deptadminrule.DeptAdminRuleResponse;
+import com.walmart.aex.sp.dto.deptadminrule.ReplItemResponse;
 import com.walmart.aex.sp.dto.mapper.DeptAdminRuleMapper;
 import com.walmart.aex.sp.entity.DeptAdminRule;
 import com.walmart.aex.sp.exception.CustomException;
@@ -16,10 +17,10 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import static com.walmart.aex.sp.util.SizeAndPackConstants.DEFAULT_MIN_REPL_ITEM_UNITS;
 import static com.walmart.aex.sp.util.SizeAndPackConstants.DEFAULT_REPL_ITEM_PIECE_RULE;
@@ -75,13 +76,13 @@ public class DeptAdminRuleServiceImpl implements DeptAdminRuleService {
     }
 
     @Override
-    public void deleteDeptAdminRules(List<DeptAdminRuleRequest> deptAdminRuleRequests) {
+    public void deleteDeptAdminRules(List<Integer> deptNbrs) {
         try {
-            if(!CollectionUtils.isEmpty(deptAdminRuleRequests)) {
+            if(!CollectionUtils.isEmpty(deptNbrs)) {
+                Set<Integer> uniqueDeptNbrs = new HashSet<>(deptNbrs);
                 List<Integer> deletionList = new ArrayList<>();
                 List<Integer> recordDoesNotExist = new ArrayList<>();
-                Set<Integer> deptNbrs = deptAdminRuleRequests.stream().map(DeptAdminRuleRequest::getDeptNbr).collect(Collectors.toSet());
-                for (Integer deptNbr: deptNbrs) {
+                for (Integer deptNbr: uniqueDeptNbrs) {
                     Optional<DeptAdminRule> existing = deptAdminRuleRepository.findById(deptNbr);
                     if(existing.isPresent()) {
                         deletionList.add(deptNbr);
@@ -89,7 +90,9 @@ public class DeptAdminRuleServiceImpl implements DeptAdminRuleService {
                         recordDoesNotExist.add(deptNbr);
                     }
                 }
-                deptAdminRuleRepository.deleteAllById(deletionList);
+                if(!CollectionUtils.isEmpty(deletionList)) {
+                    deptAdminRuleRepository.deleteAllById(deletionList);
+                }
                 if(!CollectionUtils.isEmpty(recordDoesNotExist)) {
                     log.error("Failed to delete DeptAdminRuleRequests in db. Depts: {}", recordDoesNotExist);
                 }
@@ -99,6 +102,28 @@ public class DeptAdminRuleServiceImpl implements DeptAdminRuleService {
             log.error("Failed to delete DeptAdminRuleRequests in db. Exception: {}", ex.getMessage());
             throw new CustomException("Failed to delete DeptAdminRuleRequests in db.");
         }
+    }
+
+    @Override
+    public ReplItemResponse getRepelItemRule(Long planId, Integer lvl1Nbr) {
+        ReplItemResponse response = new ReplItemResponse();
+        String plans = buyQtyProperties.getS3PlanIds();
+        int currentPlan = Math.toIntExact(planId);
+        List<Integer> s3Plans2024 = CommonUtil.getNumbersFromString(plans);
+        if(s3Plans2024.contains(currentPlan)) {
+            response.setReplItemPieceRule(buyQtyProperties.getInitialThreshold());
+            response.setMinReplItemUnits(buyQtyProperties.getReplenishmentThreshold());
+        } else {
+            List<DeptAdminRuleResponse> deptAdminRules = getDeptAdminRules(List.of(lvl1Nbr));
+            if(CollectionUtils.isEmpty(deptAdminRules)) {
+                response.setReplItemPieceRule(DEFAULT_REPL_ITEM_PIECE_RULE);
+                response.setMinReplItemUnits(DEFAULT_MIN_REPL_ITEM_UNITS);
+            } else {
+                response.setReplItemPieceRule(deptAdminRules.iterator().next().getReplItemPieceRule());
+                response.setMinReplItemUnits(deptAdminRules.iterator().next().getMinReplItemUnits());
+            }
+        }
+        return response;
     }
 
     @Override
